+++ agixt/provider/agixt.py
@@ -9,9 +9,11 @@ class AgixtProvider:
     def __init__(
         self,
         agents: list = [],
+        MAX_TOKENS: int = 16000,
         **kwargs,
     ):
         self.requirements = ["agixtsdk"]
+        self.MAX_TOKENS = int(MAX_TOKENS) if int(MAX_TOKENS) != 0 else 16000
         self.agents = ApiClient.get_agents() if agents == [] else agents
 
     async def instruct(self, prompt, tokens: int = 0):
